Further split Timing Map static methods
