+++ reamber/algorithms/timing/utils/__init__.py
@@ -1,8 +1,13 @@
 from reamber.algorithms.timing.utils.BpmChangeBase import BpmChangeBase
 from reamber.algorithms.timing.utils.BpmChangeOffset import BpmChangeOffset
 from reamber.algorithms.timing.utils.BpmChangeSnap import BpmChangeSnap
-from reamber.algorithms.timing.utils.BpmChange import BpmChange
 from reamber.algorithms.timing.utils.Snapper import Snapper
+from reamber.algorithms.timing.utils.find_lcm import find_lcm
+from reamber.algorithms.timing.utils.time_by_offset import time_by_offset
+from reamber.algorithms.timing.utils.time_by_snap import time_by_snap
+# --
 
-__all__ = [ 'BpmChangeBase', 'BpmChangeSnap', 'BpmChangeOffset', 'BpmChange',
-           'Snapper']
+from reamber.algorithms.timing.utils.BpmChange import BpmChange
+
+__all__ = ['BpmChangeBase', 'BpmChangeSnap', 'BpmChangeOffset', 'BpmChange',
+           'Snapper', 'time_by_snap', 'time_by_offset', 'find_lcm', ]
